Refactor frontend audio processing functionality
