+++ client/src/App.js
@@ -3,25 +3,12 @@ import GlobalStyles from './GlobalStyles';
 import { ThemeProvider } from 'emotion-theming';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import theme from './theme';
-import io from 'socket.io-client';
 import Container from './components/Container';
 import Notes from './pages/Notes';
 import About from './pages/About';
 import Statistics from './pages/Statistics';
-const port = process.env.SOCKET_PORT || 4000;
-
-function getSocketURL() {
-  switch (process.env.NODE_ENV) {
-    case 'development':
-      return 'localhost:' + port;
-    case 'production':
-      return undefined;
-  }
-}
 
 function App() {
-  io(getSocketURL());
-
   return (
     <ThemeProvider theme={theme}>
       <GlobalStyles />
